Initial Commit
Simple Sinatra and Sidekiq app that will clone down a repo check out the
branch from a pull request, then run pronto on the changes.

Decision to use Sidekiq so I could have control over the jobs. Just
spawning a thread was convenient, but seemed like it was a bit cowboy.